requirements.txt
psycopg2-binary
pgvector
python-dotenv
google-generativeai
sentence-transformers
numpy